Gallery2: Cleanup codes
* Since there was a lot of old code, we decided to clean it up in order to make it easier to maintain in the future
* First, replace some codes with lambda
* Second, remove all api version check because we don't need to keep compatible in lower Android version
* Third, rename some fields name to recognize easier
* Fourth, replace some deprecated methods
* Fifth, reformat codes and optimize import
* Sixth, remove duplicated Log class, use android.util.Log will better

+++ gallerycommon/src/com/android/gallery3d/common/BlobCache.java
@@ -243,7 +243,7 @@ public class BlobCache implements Closeable {
                 return false;
             }
             if (mIndexFile.length() !=
-                    INDEX_HEADER_SIZE + mMaxEntries * 12 * 2) {
+                    INDEX_HEADER_SIZE + (long) mMaxEntries * 12 * 2) {
                 Log.w(TAG, "invalid index file length");
                 return false;
             }
@@ -299,7 +299,7 @@ public class BlobCache implements Closeable {
 
     private void resetCache(int maxEntries, int maxBytes) throws IOException {
         mIndexFile.setLength(0);  // truncate to zero the index
-        mIndexFile.setLength(INDEX_HEADER_SIZE + maxEntries * 12 * 2);
+        mIndexFile.setLength(INDEX_HEADER_SIZE + (long) maxEntries * 12 * 2);
         mIndexFile.seek(0);
         byte[] buf = mIndexHeader;
         writeInt(buf, IH_MAGIC, MAGIC_INDEX_FILE);
