+++ src/com/android/gallery3d/filtershow/filters/ImageFilterEdge.java
@@ -22,6 +22,7 @@ import org.codeaurora.gallery.R;
 
 public class ImageFilterEdge extends SimpleImageFilter {
     private static final String SERIALIZATION_NAME = "EDGE";
+
     public ImageFilterEdge() {
         mName = "Edge";
     }
@@ -46,7 +47,7 @@ public class ImageFilterEdge extends SimpleImageFilter {
         int w = bitmap.getWidth();
         int h = bitmap.getHeight();
         float p = getParameters().getValue() + 101;
-        p = (float) p / 100;
+        p = p / 100;
         nativeApplyFilter(bitmap, w, h, p);
         return bitmap;
     }
